requirements.txt
Pillow==10.4.0
Requests==2.32.3
asposestorage==1.0.2
common_utils==2.0.1.dev1
dataset_utils==0.0.48
diffusers==0.29.2
einops==0.8.0
fpsample==0.3.2
google_auth==2.32.0
hid==1.0.6
imageio==2.34.1
imgaug==0.4.0
line_profiler==4.1.3
matplotlib==3.9.2
mediapipe==0.10.14
msgpack_python==0.5.6
numpy==2.0.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
plotly==5.22.0
psutil==6.0.0
pyrallis==0.3.1
scikit_learn==1.5.1
scipy==1.14.0
tabulate==0.9.0
termcolor==2.4.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.50.0
voltron_robotics==1.1.0
wandb==0.17.6
websockets==12.0
zarr==2.18.2
zerorpc==0.6.3